refactor: validate.js 파일을 검증할 값에 따라 분리
