+++ src/functions/validateNames.js
@@ -1,4 +1,3 @@
-import { MAX_VALUE_OF_ATTEMPTS } from '../constants/attemptsConstants.js';
 import {
   MAX_NAME_LENGTH,
   MAX_NUMBER_OF_PLAYERS,
@@ -6,7 +5,6 @@ import {
   MIN_NUMBER_OF_PLAYERS,
 } from '../constants/carNameConstants.js';
 import {
-  ERROR_ATTEMPTS,
   ERROR_NAME_LENGTH,
   ERROR_NUMBER_OF_PLAYERS,
 } from '../constants/messagesConstants.js';
@@ -36,26 +34,7 @@ function validateNumberOfPlayers(players) {
   }
 }
 
-export function validateNames(names) {
+export default function validateNames(names) {
   validateLengthOfNames(names);
   validateNumberOfPlayers(names);
-}
-
-function isNaturalNumber(number) {
-  const pattern = /^[1-9]\d*$/; // 1~9로 시작하는 정수 문자열
-  return pattern.test(number);
-}
-
-function isTooBig(number) {
-  return number > MAX_VALUE_OF_ATTEMPTS;
-}
-
-export function validateAttempts(attempts) {
-  if (!isNaturalNumber(attempts)) {
-    throw new Error(ERROR_ATTEMPTS);
-  }
-
-  if (isTooBig(attempts)) {
-    throw new Error(ERROR_ATTEMPTS);
-  }
 }
